# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopping Cart API",
        "description": "API documentation for Shopping Cart system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/basket": {
            "get": {
                "tags": [
                    "Basket"
                ],
                "summary": "Get current user's basket",
                "operationId": "e667cc870c8564d9f425459f56dd0d54",
                "responses": {
                    "200": {
                        "description": "User's current basket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Basket"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/basket/items": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "Add item to basket",
                "operationId": "1e43bfd55c9e6347bc827978eb719281",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddToBasketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item added to basket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BasketItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or duplicate product"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/basket/items/{id}": {
            "delete": {
                "tags": [
                    "Basket"
                ],
                "summary": "Remove item from basket",
                "operationId": "fcfa3cf2204b9117fd789b021455b730",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item removed from basket"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddToBasketRequest": {
                "title": "Add to Basket Request",
                "description": "Request payload for adding a product to the basket",
                "properties": {
                    "product_id": {
                        "description": "The ID of the product to add to the basket",
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "description": "The quantity of the product to add to the basket",
                        "type": "integer",
                        "example": 2
                    },
                    "description": {
                        "description": "Optional description for the basket item",
                        "type": "string",
                        "example": "This is a special item"
                    }
                },
                "type": "object"
            },
...
            "Basket": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "total_price": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BasketItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BasketItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "basket_id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}